Create  Activity 1 - Joins Exercise02
Create an sql that will display the following fields (attributes) – location_id, postal_code, country_id, country_name – apply a series of alias’ to the relevant tables. Then generate the output according to an INNER JOIN. With the primary key ‘location_id’ deciding what will be the prevalent table in the query.